requirements.txt
pynput
cryptography
Pillow
sounddevice
scipy
requests
pywin32; platform_system == "Windows"
